agentforge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentforge"
version = "0.1.0"
description = "企業 AI Agent 即服務平台 — YAML 定義、一行執行、三級自動修復"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "openai>=1.0",
    "google-generativeai>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
agentforge = "agentforge.cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.coverage.run]
source = ["agentforge"]

[tool.coverage.report]
show_missing = true
fail_under = 80
